refactor: LottoTicketsFactory로 생성의 책임 분리
