+++ src/test/java/lotto/domain/LottoTicketsTest.java
@@ -9,15 +9,12 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 
-//- [ ] 구입 금액을 1000으로 나눈 갯수반큼 로또 티켓을 발행한다.
-//        - [ ] 1부터 45까지의 숫자 중 중복되지 않는 6개의 숫자를 무작위로 선택한다
-
 public class LottoTicketsTest {
     @Test
     void 구입_금액으로_로또_티켓_생성() {
         PurchaseAmount purchaseAmount =new PurchaseAmount(5_000);
 
-        LottoTickets tickets = LottoTickets.create(purchaseAmount);
+        LottoTickets tickets = LottoTicketsFactory.create(purchaseAmount);
 
         assertThat(tickets.size()).isEqualTo(5);
     }
@@ -26,7 +23,7 @@ public class LottoTicketsTest {
     void 모든_로또_티켓_조회() {
         PurchaseAmount purchaseAmount =new PurchaseAmount(3_000);
 
-        LottoTickets tickets = LottoTickets.create(purchaseAmount);
+        LottoTickets tickets = LottoTicketsFactory.create(purchaseAmount);
 
         List<Lotto> lottos = tickets.getLottos();
         assertThat(lottos).hasSize(3);
@@ -42,7 +39,7 @@ public class LottoTicketsTest {
         Lotto manual2 = new Lotto(7, 8, 9, 10, 11, 12);
         ManualLottos manualLottos = new ManualLottos(Arrays.asList(manual1, manual2));
 
-        LottoTickets tickets = LottoTickets.create(purchaseAmount, manualLottos);
+        LottoTickets tickets = LottoTicketsFactory.create(purchaseAmount, manualLottos);
 
         assertThat(tickets.size()).isEqualTo(2);
         assertThat(tickets.getManualCount()).isEqualTo(2);
@@ -56,7 +53,7 @@ public class LottoTicketsTest {
         Lotto manual2 = new Lotto(7, 8, 9, 10, 11, 12);
         ManualLottos manualLottos = new ManualLottos(Arrays.asList(manual1, manual2));
 
-        LottoTickets tickets = LottoTickets.create(purchaseAmount, manualLottos);
+        LottoTickets tickets = LottoTicketsFactory.create(purchaseAmount, manualLottos);
 
         assertThat(tickets.size()).isEqualTo(5);
         assertThat(tickets.getManualCount()).isEqualTo(2);
@@ -71,7 +68,7 @@ public class LottoTicketsTest {
         Lotto manual3 = new Lotto(13, 14, 15, 16, 17, 18);
         ManualLottos manualLottos = new ManualLottos(Arrays.asList(manual1, manual2, manual3));
 
-        assertThatThrownBy(() -> LottoTickets.create(purchaseAmount, manualLottos))
+        assertThatThrownBy(() -> LottoTicketsFactory.create(purchaseAmount, manualLottos))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining("수동로또 개수가 구입 가능한 개수보다 많습니다.");
     }
